chore(stats): move cache stats to cache package
